Add lab2xyz "UserWarning: Color data out of range: Z < 0" example
Sometimes during conversion from L*a*b* to xyz color space a user warning is raised if the values are invalid.
This commit adds an example (L*a*b* values [0, -13.70470524, 27.88690567]) of when this happens.

"UserWarning: Color data out of range: Z < 0 in 1 pixels"

This subject was discussed in https://github.com/scikit-image/scikit-image/issues/4506